requirements.txt
about-time==4.2.1
alive-progress==3.1.5
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
dask==2024.8.1
faiss-cpu==
filelock==3.15.4
fsspec==2024.6.1
grapheme==0.6.0
huggingface-hub==0.24.6
idna==3.8
importlib_metadata==8.4.0
ipykernel
Jinja2==3.1.4
joblib==1.4.2
locket==1.0.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.0
packaging==24.1
pandas==2.2.2
partd==1.4.2
pyarrow==17.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.1
six==1.16.0
sympy==1.13.2
threadpoolctl==3.5.0
tokenizers==0.19.1
toolz==0.12.1
torch==2.4.0
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
zipp==3.20.1
# numpy > 2 not compatible with torch
# torch==1.11.0